Refactor PDE headers in preparation for installation
