+++ src/equation_systems/PDEBase.H
@@ -4,12 +4,10 @@
 #include <string>
 
 #include "Factory.H"
-#include "incflo_enums.H"
-#include "FieldDescTypes.H"
-#include "FieldRepo.H"
-#include "PDEHelpers.H"
-#include "SimTime.H"
 #include "CollMgr.H"
+#include "PDEFields.H"
+#include "FieldDescTypes.H"
+#include "incflo_enums.H"
 
 namespace amr_wind {
