+++ src/equation_systems/PDEBase.cpp
@@ -1,11 +1,23 @@
 #include "PDEBase.H"
 #include "SchemeTraits.H"
 #include "CFDSim.H"
+#include "FieldRepo.H"
+#include "PDEHelpers.H"
+#include "incflo_enums.H"
 
 #include "AMReX_ParmParse.H"
 
 namespace amr_wind {
 namespace pde {
+
+PDEFields::PDEFields(FieldRepo& repo_in, const std::string& var_name)
+    : repo(repo_in)
+    , field(repo.get_field(var_name))
+    , mueff(repo.get_field(pde_impl::mueff_name(var_name)))
+    , src_term(repo.get_field(pde_impl::src_term_name(var_name)))
+    , diff_term(repo.get_field(pde_impl::diff_term_name(var_name)))
+    , conv_term(repo.get_field(pde_impl::conv_term_name(var_name)))
+{}
 
 PDEMgr::PDEMgr(CFDSim& sim)
     : m_sim(sim)
